+++ tslearn/tests/test_svm.py
@@ -25,8 +25,23 @@ def test_gamma_value_svm():
                                                      fit_time=True)
 
         cdist_mat = cdist_gak(time_series, sigma=np.sqrt(gamma / 2.))
-
         np.testing.assert_allclose(sklearn_X, cdist_mat)
+
+    # Invalid gamma 0 for gak kernel
+    for cls in [TimeSeriesSVC, TimeSeriesSVR]:
+        estimator = cls(kernel="gak", gamma=0)
+        with pytest.raises(RuntimeError):
+            estimator.fit(time_series, labels)
+
+    # Invalid computed gamma for gak kernel
+    X = ([np.ones(3), np.ones(3)])
+    y = [0, 1]
+
+    for cls in [TimeSeriesSVC, TimeSeriesSVR]:
+        estimator = cls(kernel="gak")
+        with pytest.raises(RuntimeError):
+            estimator.fit(X, y)
+
 
 def test_attributes():
     n, sz, d = 5, 10, 3